long id eliminated

+++ src/main/java/solutions/dmitrikonnov/einstufungstest/domainlayer/ETAntwortBogenDto.java
@@ -13,6 +13,6 @@ import java.util.Map;
 public class ETAntwortBogenDto implements Serializable {
 
     private static final long serialVersionUID = -1798070786993154676L;
-    private final Long antwortBogenId;
+    private final Integer antwortBogenId;
     private final Map<Integer, ArrayList<String>> itemHashZuAntwortMap;
 }
